third_party/WebKit/LayoutTests/inspector-protocol/timeline/timeline-timer.html

<html>
<head>
<script type="text/javascript" src="../../http/tests/inspector-protocol/inspector-protocol-test.js"></script>
<script type="text/javascript" src="../../http/tests/inspector-protocol/tracing-test.js"></script>
<script>

function performActions()
{
    var callback;
    var promise = new Promise((fulfill) => callback = fulfill);
    var timerId = setTimeout(function()
    {
        evaluateInFrontend("InspectorTest.testFunctionTimerFired(" + timerId + ", " + timerId2 + ")");
        callback();
    }, 0);

    var timerId2 = setTimeout(function() { }, 0);
    clearTimeout(timerId2);
    return promise;
}

function test()
{
    InspectorTest.invokeAsyncWithTracing("performActions", finish);

    var firedTimerId;
    var removedTimerId;
    InspectorTest.testFunctionTimerFired = function(timerId1, timerId2)
    {
        firedTimerId = timerId1;
        removedTimerId = timerId2;
        InspectorTest.log("SUCCESS: testFunctionTimerFired");
    }

    function finish(devtoolsEvents)
    {
        function hasTimerId(id, e) { return e.args.data.timerId === id}

        var installTimer1 = InspectorTest.findEvent("TimerInstall", "I", hasTimerId.bind(this, firedTimerId));
        var installTimer2 = InspectorTest.findEvent("TimerInstall", "I", hasTimerId.bind(this, removedTimerId));

        InspectorTest.log("TimerInstall has frame: " + !!installTimer1.args.data.frame);
        InspectorTest.log("TimerInstall frames match: " + (installTimer1.args.data.frame === installTimer2.args.data.frame));

        InspectorTest.findEvent("TimerRemove", "I", hasTimerId.bind(this, removedTimerId));
        InspectorTest.findEvent("TimerFire", "X", hasTimerId.bind(this, firedTimerId));

        InspectorTest.log("SUCCESS: found all expected events.");
        InspectorTest.completeTest();
    }
}
</script>
</head>
<body onLoad="runTest();">
<div id="myDiv">DIV</div>
</body>
</html>
