third_party/WebKit/LayoutTests/inspector-protocol/timeline/timeline-timer-expected.txt

DIV
Recording started
SUCCESS: testFunctionTimerFired
Tracing complete
TimerInstall has frame: true
TimerInstall frames match: true
SUCCESS: found all expected events.
